lib/vfscore: Adopt responsibility for file VMAs
This change moves the implementation of file-backed VMAs from ukvmem to
vfscore, freeing the former from an unorthogonal responsibility.
Calling code in posix-mmap is updated.

Checkpatch-Ignore: BIT_MACRO
Checkpatch-Ignore: PARENTHESIS_ALIGNMENT
Checkpatch-Ignore: LONG_LINE
