+++ lib/vfscore/Makefile.uk
@@ -17,6 +17,12 @@ LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/lookup.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/fops.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/subr_uio.c
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/extra.ld
+ifeq ($(CONFIG_LIBUKVMEM),y)
+LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/vma_file.c|isr
+ifneq ($(filter y,$(CONFIG_LIBUKVMEM_TEST) $(CONFIG_LIBUKTEST_ALL)),)
+LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/tests/test_file_vma.c
+endif
+endif
 ifneq ($(filter y,$(CONFIG_LIBVFSCORE_AUTOMOUNT) \
 		  $(CONFIG_LIBVFSCORE_AUTOUNMOUNT)),)
 LIBVFSCORE_SRCS-y += $(LIBVFSCORE_BASE)/automount.c
