+++ lib/vfscore/exportsyms.uk
@@ -164,3 +164,4 @@ vfscore_fstat
 vfscore_fchmod
 vfscore_fcntl
 vfscore_ioctl
+uk_vma_file_ops
